ml-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.36.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
pandas==2.1.4
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
ultralytics==8.0.196
clip-by-openai==1.0
spacy==3.7.2
nltk==3.8.1
openai==1.3.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tensorflow==2.15.0
keras==2.15.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
redis==5.0.1
celery==5.3.4
boto3==1.34.0
Cython==3.0.6 